requirements.txt
contextlib2==0.5.4
dj-database-url==0.4.1
Django==1.10.2
django-appconf==1.0.2
django-compressor==2.1
django-configurations==2.0
django-storages-redux==1.3.2
easy-thumbnails==2.3
Pillow==3.4.0
psycopg2==2.6.2
raven==5.27.1
rcssmin==1.0.6
rjsmin==1.0.12
